--- SkoolOS/logs/skoolos_06162020-221844
@@ -1,55 +1,0 @@
-Start time: Tuesday, June 16, 2020 22:18:48
-
-Event: Opened file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/run.py
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Created file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice.run.html
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Opened file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice.run.html
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-wrote bgservice.run.html
-Event: Wrote to a file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice.run.html
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Opened file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/__pycache__/test.cpython-38.pyc
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Opened file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/test.py
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Created file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/__pycache__/test.cpython-38.pyc.139630150929712
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Opened file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/__pycache__/test.cpython-38.pyc.139630150929712
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Wrote to a file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/__pycache__/test.cpython-38.pyc.139630150929712
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Moved a file out
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/__pycache__/test.cpython-38.pyc.139630150929712
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Moved a file in
-Event Path: /home/nkenschaft/Sysadmin/skoolos/bgservice/__pycache__/test.cpython-38.pyc
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Created file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/SkoolOS/logs/skoolos_06162020-221848
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
-Event: Opened file
-Event Path: /home/nkenschaft/Sysadmin/skoolos/SkoolOS/logs/skoolos_06162020-221848
-Timestamp: Tuesday, June 16, 2020 22:18:48
-
